+++ langs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 ## Unreleased
+
+## 0.9.0 - 2020-12-16
+- Updated `javascript` and `typescript` grammars.
+
+## 0.8.0 - 2020-12-15
+- Added `elm`.
+- Handled underscores in language names correctly.  This enabled using `c-sharp` instead of `c_sharp`.
+
 ## 0.7.2 - 2020-10-03
 - Set `tree-sitter-hl-default-patterns` to nil if there is no query file, instead of signaling an error.